backbone_extraction/requirements.txt
beautifulsoup4 == 4.11.1
DateTime == 4.7
igraph == 0.10.2
multiprocess == 0.70.14
networkx == 2.8.8
nltk == 3.7
numpy == 1.23.5
pandas == 1.5.2
pymongo == 4.3.3
python-dotenv == 0.21.0
tqdm == 4.64.1
regex == 2022.10.31
wordcloud == 1.8.2.2
lxml == 4.9.2
stopwords == 1.0.0